kvx qsfp: Remove signed to unsigned cast
Summary:
kvx_qsfp_transceiver_compliance_code returns a u8, but could return
-ETIMEDOUT if the EEPROM wasn't/couldn't be read yet. Since the caller
wasn't checking the returned value (and probably couldn't anyway), this
could result in an incorrect subset of modes being reported as supported
by some QSFP modules (Q28-PC02 in particular).

Get rid of kvx_qsfp_transceiver_compliance_code, and report all modes as
being supported when the EEPROM info isn't available.

Ref T22393

Test Plan: CI

Reviewers: O39 ethernet, O129 linux drivers

Maniphest Tasks: T22393

Differential Revision: https://phab.kalray.eu/D34603

+++ drivers/net/ethernet/kalray/kvx-qsfp.h
@@ -438,8 +438,8 @@ int kvx_qsfp_set_eeprom(struct kvx_qsfp *qsfp, struct ethtool_eeprom *ee, u8 *da
 bool is_cable_connected(struct kvx_qsfp *qsfp);
 bool is_cable_copper(struct kvx_qsfp *qsfp);
 void kvx_qsfp_parse_support(struct kvx_qsfp *qsfp, unsigned long *support);
-u8 kvx_qsfp_transceiver_id(struct kvx_qsfp *qsfp);
-u32 kvx_qsfp_transceiver_nominal_br(struct kvx_qsfp *qsfp);
+int kvx_qsfp_transceiver_id(struct kvx_qsfp *qsfp);
+int kvx_qsfp_transceiver_nominal_br(struct kvx_qsfp *qsfp);
 int kvx_qsfp_ops_register(struct kvx_qsfp *qsfp, struct kvx_qsfp_ops *ops, void *ops_data);
 bool kvx_qsfp_int_flags_supported(struct kvx_qsfp *qsfp);
